Added 'CPH' parameter to expression_filter. May or may not be helpful.
Also, the output of expression filter is a two element list. 'Result' is a data frame showing the filtered data and a column with whatever statistic the user specified to filter it by. This is to visualize what features had what filter/rank statistic. 'dat' is the data or if specified the DGEList that is returned for further analysis

Added 'sep' parameter to plot_pca because moving columns to rownames and back sometimes replaces certain separators with others. If this happens the sample names can be repaired to match that of metadata. This fixes errors when trying to join with metadata by sample name.